// src/api/generated/model/revenueReportEntryEntryCategory.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * MiniCRM API
 * API Documentation for MiniCRM Application
 * OpenAPI spec version: 1.0.0
 */

/**
 * Category of the revenue entry (REGULAR, FLAT_RATE, WEEKLY)
 */
export type RevenueReportEntryEntryCategory = typeof RevenueReportEntryEntryCategory[keyof typeof RevenueReportEntryEntryCategory];


export const RevenueReportEntryEntryCategory = {
  REGULAR: 'REGULAR',
  FLAT_RATE: 'FLAT_RATE',
  WEEKLY: 'WEEKLY',
} as const;